add scholarlogo and sidelink components to navbar
